# blockchain: Optimize stake and vote lookups.