The functions that are used to keep stake version and vote versions humming along are relatively expensive due to repeated chain walks. This commit removes the most egregious sinners by memoizing the answer. It should be noted that we can not memoize the walk to the prior interval because that would require one to memoize all blocks up to the prior interval.  This is in practice not a big issue because relatively new `blockNodes` are in memory.  The real savings happen when a block does not need to be loaded from disk.

Here are some figures from a quad core Xeon 2.66GHz:

> stakeversion_test.go:930: running: shallow no cache with 20160 nodes
> stakeversion_test.go:1000:   run time (setup 0) 712.273µs
> stakeversion_test.go:1000:   run time (run 1) 576.898µs
> stakeversion_test.go:1000:   run time (run 2) 571.683µs
> stakeversion_test.go:1000:   run time (run 3) 571.424µs
> stakeversion_test.go:1000:   run time (run 4) 570.918µs
> stakeversion_test.go:930: running: shallow cache with 20160 nodes
> stakeversion_test.go:1000:   run time (setup 0) 775.044µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 1) 74.316µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 2) 57.283µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 3) 57.235µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 4) 57.068µs memory cost: 116
> stakeversion_test.go:930: running: deep no cache with 201600 nodes
> stakeversion_test.go:1000:   run time (setup 0) 700.987µs
> stakeversion_test.go:1000:   run time (run 1) 571.4µs
> stakeversion_test.go:1000:   run time (run 2) 569.362µs
> stakeversion_test.go:1000:   run time (run 3) 569.154µs
> stakeversion_test.go:1000:   run time (run 4) 569.319µs
> stakeversion_test.go:930: running: deep cache with 201600 nodes
> stakeversion_test.go:1000:   run time (setup 0) 723.686µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 1) 57.565µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 2) 55.698µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 3) 55.478µs memory cost: 116
> stakeversion_test.go:1000:   run time (run 4) 56.046µs memory cost: 116